Create and implement EntryPolicy rules

+++ app/Http/Controllers/EntryController.php
@@ -36,20 +36,22 @@ class EntryController extends Controller
      */
     public function store(Request $request, Journal $journal)
     {
-        $request->validate([
-            'title' => 'nullable|max:512'
-        ]);
+        if (Auth::user()->can('create', Entry::class)) {
+            $request->validate([
+                'title' => 'nullable|max:512'
+            ]);
 
-        $entry = new Entry;
-        $entry->title = $request->title ? $request->title : '[Untitled]';
-        $entry->body = $request->body;
-        $entry->status = 'draft';
-        $entry->author()->associate(Auth::user());
+            $entry = new Entry;
+            $entry->title = $request->title ? $request->title : '[Untitled]';
+            $entry->body = $request->body;
+            $entry->status = 'draft';
+            $entry->author()->associate(Auth::user());
 
-        $journal = Journal::find($request->journal_id);
-        $journal->entries()->save($entry);
+            $journal = Journal::find($request->journal_id);
+            $journal->entries()->save($entry);
 
-        $request->session()->flash('status', "<strong>{$entry->title}</strong> has been saved.");
+            $request->session()->flash('status', "<strong>{$entry->title}</strong> has been saved.");
+        }
         return redirect()->route('journal.show', compact('journal'));
     }
 
@@ -62,10 +64,19 @@ class EntryController extends Controller
      */
     public function show(Entry $entry, Journal $journal)
     {
-        // TODO: Validate request
-        $nextEntry = $journal->getEntryAfter($entry);
-        $previousEntry = $journal->getEntryBefore($entry);
-        return view('entry.show', compact('entry', 'journal', 'nextEntry', 'previousEntry'));
+        if (Auth::user()->can('view', $entry)) {
+            $nextEntry = $journal->getEntryAfter($entry);
+            $previousEntry = $journal->getEntryBefore($entry);
+            return view('entry.show', compact('entry', 'journal', 'nextEntry', 'previousEntry'));
+        }
+
+        // Show a flash message if the user belongs to the journal.
+        if (Auth::user()->isInJournal($journal)) {
+            $request->session()->flash('status', "{$journal->current_user->name} has <strong>{$journal->title}</strong> right now. You'll be able to read it when it's your turn.");
+        }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -77,10 +88,12 @@ class EntryController extends Controller
      */
     public function edit(Entry $entry, Journal $journal)
     {
-
-        if ($entry->status == 'draft') {
+        if (Auth::user()->can('update', $entry)) {
             return view('entry.edit', compact('entry', 'journal'));
         }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -93,15 +106,20 @@ class EntryController extends Controller
      */
     public function update(Request $request, Entry $entry, Journal $journal)
     {
-        $request->validate([
-            'title' => 'nullable|max:512'
-        ]);
-        $entry->body = $request->body;
-        $entry->title = $request->title;
-        $entry->save();
+        if (Auth::user()->can('update', $entry)) {
+            $request->validate([
+                'title' => 'nullable|max:512'
+            ]);
+            $entry->body = $request->body;
+            $entry->title = $request->title;
+            $entry->save();
 
-        $request->session()->flash('status', "<strong>{$entry->title}</strong> has been saved.");
-        return redirect()->route('journal.show', compact('journal'));
+            $request->session()->flash('status', "<strong>{$entry->title}</strong> has been saved.");
+            return redirect()->route('journal.show', compact('journal'));
+        }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -114,15 +132,18 @@ class EntryController extends Controller
      */
     public function destroy(Request $request, Entry $entry, Journal $journal)
     {
-        $entry->delete();
-        $request->session()->flash('status',
-            "<strong>{$entry->title}</strong> has been deleted.
-            <a href=\"#\" onclick=\"event.preventDefault(); document.getElementById('undo-form').submit();\">Undo</a>
-            <form id=\"undo-form\" class=\"d-none\" method=\"post\" action='/entry/undodelete'>
-                <input type='hidden' name='_token' value='" . csrf_token() . "'>
-                <input type='hidden' name='entry_id' value='{$entry->id}'>
-            </form>"
-        );
+        if (Auth::user()->can('delete', $entry)) {
+            $entry->delete();
+            $request->session()->flash('status',
+                "<strong>{$entry->title}</strong> has been deleted.
+                <a href=\"#\" onclick=\"event.preventDefault(); document.getElementById('undo-form').submit();\">Undo</a>
+                <form id=\"undo-form\" class=\"d-none\" method=\"post\" action='/entry/undodelete'>
+                    <input type='hidden' name='_token' value='" . csrf_token() . "'>
+                    <input type='hidden' name='entry_id' value='{$entry->id}'>
+                </form>"
+            );
+        }
+
         return redirect()->route('journal.show', compact('journal'));
     }
 
@@ -134,12 +155,13 @@ class EntryController extends Controller
      */
     public function undoDelete(Request $request)
     {
-        // TODO: Validate request
+        if (Auth::user()->can('restore', $entry)) {
+            $entry = Entry::withTrashed()->find($request->input('entry_id'));
+            $entry->restore();
+            $request->session()->flash('status', "<strong>{$entry->title}</strong> has been restored.");
+            $journal = $entry->journal;
+        }
 
-        $entry = Entry::withTrashed()->find($request->input('entry_id'));
-        $entry->restore();
-        $request->session()->flash('status', "<strong>{$entry->title}</strong> has been restored.");
-        $journal = $entry->journal;
         return redirect()->route('journal.show', compact('journal'));
     }
 }
